update module launch welcome page
